# ledger-service/openapi.yaml
openapi: 3.1.0
info:
  title: Payments Ledger API
  version: 0.1.0
  description: >
    Double-entry payments ledger with exactly-once money movement. All amounts
    are exact decimal values in a single currency per transaction. Money movement
    requires an Idempotency-Key header so requests are safe to retry.
servers:
  - url: http://localhost:8080
paths:
  /v1/accounts:
    post:
      summary: Open an account
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAccountResponse'
        '400':
          $ref: '#/components/responses/Error'
  /v1/accounts/{accountId}/balance:
    get:
      summary: Read an account balance
      operationId: getBalance
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Current balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '404':
          $ref: '#/components/responses/Error'
  /v1/transactions:
    post:
      summary: Post a balanced double-entry transaction
      operationId: postTransaction
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostTransactionRequest'
      responses:
        '201':
          description: Transaction posted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostTransactionResponse'
        '400':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
components:
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    CreateAccountRequest:
      type: object
      required: [currency]
      properties:
        currency:
          type: string
          examples: [USD]
        allowOverdraft:
          type: boolean
          default: false
    CreateAccountResponse:
      type: object
      properties:
        accountId:
          type: string
          format: uuid
        currency:
          type: string
        status:
          type: string
    BalanceResponse:
      type: object
      properties:
        accountId:
          type: string
          format: uuid
        amount:
          type: string
          examples: ["90.00"]
        currency:
          type: string
    PostTransactionRequest:
      type: object
      required: [postings]
      properties:
        postings:
          type: array
          minItems: 2
          items:
            $ref: '#/components/schemas/Posting'
        metadata:
          type: object
          additionalProperties:
            type: string
    Posting:
      type: object
      required: [accountId, direction, amount, currency]
      properties:
        accountId:
          type: string
          format: uuid
        direction:
          type: string
          enum: [DEBIT, CREDIT]
        amount:
          type: string
          examples: ["10.00"]
        currency:
          type: string
    PostTransactionResponse:
      type: object
      properties:
        transactionId:
          type: string
          format: uuid
        status:
          type: string
          examples: [POSTED]
        balances:
          type: array
          items:
            $ref: '#/components/schemas/BalanceResponse'
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        correlationId:
          type: string
        timestamp:
          type: string
          format: date-time